refactor: relocate content data
