+++ src/components/introduce/IntroContents.tsx
@@ -1,8 +1,7 @@
 import styled from '@emotion/styled';
 import MailIcon from 'assets/MailIcon';
 import StackButton from 'components/common/StackButton';
-
-import INTRODUCE_DATA from './introduceData';
+import INTRODUCE_DATA from 'content/introduceData';
 
 const IntroContents = () => {
   return (
